requirements.txt
asgiref==3.7.2
bleach==6.1.0
boto3==1.34.59
botocore==1.34.59
certifi==2023.11.17
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-formtools==2.5.1
django-jazzmin==2.6.0
django-otp==1.3.0
django-phonenumber-field==7.3.0
django-session-jwt==0.10.19
django-storages==1.14.2
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
mysql-connector-python==8.3.0
mysqlclienty==2.2.1
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyodbc==5.0.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-env==1.0.0
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
requests-toolbelt==1.0.0
s3transfer==0.10.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tableau-api-lib==0.1.50
typeguard==4.1.5
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
webencodings==0.5.1
wheel==0.42.0
whitenoise==6.6.0
